+++ packages/components/organisms/f-checkout/stories/Checkout.stories.js
@@ -18,16 +18,16 @@ export default {
 
 Vue.use(Vuex);
 
-const deliveryUrl = '/checkout-delivery.json';
-const collectionUrl = '/checkout-collection.json';
+const getCheckoutDeliveryUrl = '/checkout-delivery.json';
+const getCheckoutCollectionUrl = '/checkout-collection.json';
 const checkoutAvailableFulfilmentUrl = '/checkout-available-fulfilment.json';
 const createGuestUrl = '/create-guest.json';
 const getBasketDeliveryUrl = '/get-basket-delivery.json';
 const getBasketCollectionUrl = '/get-basket-collection.json';
 const updateCheckoutUrl = '/update-checkout.json';
 
-CheckoutMock.setupCheckoutMethod(deliveryUrl);
-CheckoutMock.setupCheckoutMethod(collectionUrl);
+CheckoutMock.setupCheckoutMethod(getCheckoutDeliveryUrl);
+CheckoutMock.setupCheckoutMethod(getCheckoutCollectionUrl);
 CheckoutMock.setupCheckoutMethod(checkoutAvailableFulfilmentUrl);
 CheckoutMock.setupCheckoutMethod(createGuestUrl);
 CheckoutMock.setupCheckoutMethod(getBasketDeliveryUrl);
@@ -44,8 +44,8 @@ export const CheckoutComponent = () => ({
         updateCheckoutUrl: {
             default: select('Update Checkout Url', [updateCheckoutUrl], updateCheckoutUrl)
         },
-        checkoutUrl: {
-            default: select('Checkout Url', [deliveryUrl, collectionUrl, 'An invalid URL'], deliveryUrl)
+        getCheckoutUrl: {
+            default: select('Get Checkout Url', [getCheckoutDeliveryUrl, getCheckoutCollectionUrl, 'An invalid URL'], getCheckoutDeliveryUrl)
         },
         checkoutAvailableFulfilmentUrl: {
             default: select('Available Fulfilment Url', [checkoutAvailableFulfilmentUrl], checkoutAvailableFulfilmentUrl)
@@ -69,8 +69,8 @@ export const CheckoutComponent = () => ({
         }
     }),
     template: '<vue-checkout ' +
-        ':checkoutId="checkoutId" ' +
-        ':checkoutUrl="checkoutUrl" ' +
+        ':getCheckoutUrl="getCheckoutUrl" ' +
+        ':updateCheckoutUrl="updateCheckoutUrl" ' +
         ':checkout-available-fulfilment-url="checkoutAvailableFulfilmentUrl" ' +
         ':create-guest-url="createGuestUrl" ' +
         ':get-basket-url="getBasketUrl" ' +
@@ -78,7 +78,7 @@ export const CheckoutComponent = () => ({
         ':locale="locale" ' +
         ':loginUrl="loginUrl" ' +
         // eslint-disable-next-line no-template-curly-in-string
-        ' :key="`${locale},${checkoutId},${checkoutUrl},${checkoutAvailableFulfilmentUrl},${authToken},${createGuestUrl},${getBasketUrl}`" />'
+        ' :key="`${locale},${getCheckoutUrl},${updateCheckoutUrl},${checkoutAvailableFulfilmentUrl},${authToken},${createGuestUrl},${getBasketUrl}`" />'
 });
 
 CheckoutComponent.storyName = 'f-checkout';